requirements.txt
altgraph==0.16.1
amqp==2.5.1
auto-py-to-exe==2.6.5
beautifulsoup4==4.7.1
billiard==3.6.1.0
bottle==0.12.17
bottle-websocket==0.2.9
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
Cython==0.29.17
chardet==3.0.4
Click==7.0
comtypes==1.1.7
cycler==0.10.0
Django==2.2
django-bootstrap3==11.1.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.14
Eel==0.10.4
face-recognition-models==0.3.0
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
Gooey==1.0.2
greenlet==0.4.15
idna==2.8
importlib-metadata==0.23
imutils==0.5.3
joblib==0.14.1
Js2Py==0.68
Kivy==1.10.1
Kivy-Garden==0.1.4
kivy.deps.angle==0.1.7
kivy.deps.glew==0.1.10
kivy.deps.gstreamer==0.1.13
kivy.deps.sdl2==0.1.18
kivy.deps.sdl2-dev==0.1.18
kiwisolver==1.0.1
kombu==4.6.5
macholib==1.11
matplotlib==3.0.3
more-itertools==7.2.0
MouseInfo==0.0.4
mysql-connector==2.2.9
numpy==1.16.2
opencv-contrib-python==4.2.0.32
packaging==20.3
pandas==0.25.0
pefile==2019.4.18
Pillow==5.4.1
Pillow-PIL==0.1.dev0
pillowcover-gui==1.2.0
pipwin==0.4.7
plyer==1.4.2
prettytable==0.7.2
psutil==5.6.1
PyAutoGUI==0.9.47
pycparser==2.19
pygame==1.9.4
PyGetWindow==0.0.6
Pygments==2.3.1
PyInstaller==3.4
pyjsparser==2.7.1
PyMsgBox==1.0.7
pyparsing==2.3.1
pyperclip==1.7.0
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.22
python-dateutil==2.8.0
python-decouple==3.1
pyttsx3==2.81
PyTweening==1.0.3
pytz==2019.1
pywin32==227
pywin32-ctypes==0.2.0
requests==2.21.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
SpeechRecognition==3.8.1
SQLite3Database==0.2.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.24.1
vine==1.3.0
whichcraft==0.6.0
wikipedia==1.4.0
word2number==1.1
wxPython==4.0.4
yieldfrom.urllib.request==0.1.1
zipp==0.6.0
